refactor(order-sync): wrap up
