+++ src/index.js
@@ -108,4 +108,4 @@ export { default as createSyncCategories } from './sync/categories'
 export { default as createSyncCustomers } from './sync/customers'
 export { default as createSyncInventories } from './sync/inventories'
 export { default as createSyncProducts } from './sync/products'
-export { default as createSyncOrder } from './sync/order'
+export { default as createSyncOrders } from './sync/orders'
