+++ test/sync/order-sync.spec.js
@@ -1,19 +1,18 @@
 import test from 'tape'
-import ordersSyncFn, { actionGroups } from '../../src/sync/order'
+import orderSyncFn, { actionGroups } from '../../src/sync/orders'
 import {
   baseActionsList,
-  referenceActionsList,
 } from '../../src/sync/order-actions'
 
-test.only('Sync::order', (t) => {
-  let ordersSync
+test('Sync::order', (t) => {
+  let orderSync
   function setup () {
-    ordersSync = ordersSyncFn()
+    orderSync = orderSyncFn()
   }
 
   t.test('should export action group list', (t) => {
     t.deepEqual(actionGroups, [
-      'base', 'references',
+      'base',
     ])
     t.end()
   })
@@ -22,68 +21,36 @@ test.only('Sync::order', (t) => {
     t.deepEqual(
       baseActionsList,
       [
-        { action: 'setShipmentState', key: 'shipmentState' },
-        { action: 'setPaymentState', key: 'paymentState' },
-        { action: 'setOrderState', key: 'orderState' },
+        { action: 'changeOrderState', key: 'orderState' },
+        { action: 'changePaymentState', key: 'paymentState' },
+        { action: 'changeShipmentState', key: 'shipmentState' },
       ],
       'correctly define base actions list'
     )
 
-    t.deepEqual(
-      referenceActionsList,
-      [],
-      'correctly define reference actions list'
-    )
-
     t.end()
   })
 
-  t.test('should build `setShipmentState` action', (t) => {
-    setup()
-
-    const before = {
-      shipmentState: 'Shipped',
-    }
-    const now = {
-      shipmentState: 'Ready',
-    }
-
-    const actual = ordersSync.buildActions(now, before)
-    const expected = [{ action: 'setShipmentState', shipmentState: 'Ready' }]
-    t.deepEqual(actual, expected)
-
-    t.end()
-  })
-
-  t.test('should build `setPaymentState` action', (t) => {
-    setup()
-
-    const before = {
-      paymentState: 'BalanceDue',
-    }
-    const now = {
-      paymentState: 'Paid',
-    }
-
-    const actual = ordersSync.buildActions(now, before)
-    const expected = [{ action: 'setPaymentState', paymentState: 'Paid' }]
-    t.deepEqual(actual, expected)
-
-    t.end()
-  })
-
-  t.test('should build `setOrderState` action', (t) => {
+  t.test('should build *state actions', (t) => {
     setup()
 
     const before = {
       orderState: 'Open',
+      paymentState: 'Pending',
+      shipmentState: 'Ready',
     }
     const now = {
-      orderState: 'Confirmed',
+      orderState: 'Complete',
+      paymentState: 'Paid',
+      shipmentState: 'Shipped',
     }
 
-    const actual = ordersSync.buildActions(now, before)
-    const expected = [{ action: 'setOrderState', orderState: 'Confirmed' }]
+    const actual = orderSync.buildActions(now, before)
+    const expected = [
+      { action: 'changeOrderState', orderState: 'Complete' },
+      { action: 'changePaymentState', paymentState: 'Paid' },
+      { action: 'changeShipmentState', shipmentState: 'Shipped' },
+    ]
     t.deepEqual(actual, expected)
 
     t.end()
